requirements.txt
Django==1.11
MySQL-python==1.2.5
mysqlclient==1.3.10
django-webpack-loader==0.4.1
djangorestframework==3.6.2
django-extensions-1.7.8
nltk-3.2.2
textblob-0.12.0
django-debug-toolbar-1.8
lxml-3.8.0
xml-0.2.13
httplib2-0.10.3
numpy-1.13.0